第4次重构代码
